types: extract fragment types to fragment.ts
